Added basic jump & other misc tweaks
Added "Platform" tag
Added collision detectors that would help determine whether or not the player was currently touching an object with the "Platform" tag.
Decreased the Y value for gravity from -9.81 to -30. This will result in harsher gravity, allowing for faster falls when jumping. Jump Force has been increased to compensate.

TODO:
Allow the jump to be shorter if the player only taps the jump button, and vice versa.
Add a collider to the bottom of the players feet to act as the new floor detector to prevent possible unintended wall jumping.